+++ app/src/main/java/com/luisenricke/simpleroomapp/data/dao/MedicineDAO.kt
@@ -13,7 +13,7 @@ abstract class MedicineDAO : Base<Medicine>,
     Base.PrimaryKeyDAO<Medicine> {
 
     @Query("SELECT COUNT(*) FROM ${SCHEMA.TABLE}")
-    abstract override fun count(): Int
+    abstract override fun count(): Long
 
     @Query("SELECT * FROM ${SCHEMA.TABLE}")
     abstract override fun get(): List<Medicine>
@@ -22,14 +22,14 @@ abstract class MedicineDAO : Base<Medicine>,
     abstract override fun drop()
 
     @Query("SELECT * FROM ${SCHEMA.TABLE} WHERE id = :id")
-    abstract override fun get(id: Int): Medicine
+    abstract override fun get(id: Long): Medicine
 
     @Query("SELECT * FROM ${SCHEMA.TABLE} WHERE id IN(:ids)")
-    abstract override fun get(ids: IntArray): List<Medicine>
+    abstract override fun get(ids: LongArray): List<Medicine>
 
     @Query("DELETE FROM ${SCHEMA.TABLE} WHERE id = :id")
-    abstract override fun delete(id: Int): Int
+    abstract override fun delete(id: Long): Int
 
     @Query("DELETE FROM ${SCHEMA.TABLE} WHERE id IN(:ids)")
-    abstract override fun deletes(ids: IntArray): Int
+    abstract override fun deletes(ids: LongArray): Int
 }
